Added 2525C symbols
Added 2525C symbology. For testing, the app will display a 2525C symbol
when the map is tapped. This test will be removed, but the underlying
functionality is likely permanent: check for the symbol library; copy it
into place if it is absent; and set up a MessageGroupLayer based on the
symbol library.