Fixing spot reports
This may resolve the spot report part of #117. (@jrweakland, will you
please try it and let me know?)

Using map's last known spatial reference if the map's current spatial
reference is null. My theory is that on some devices, the map's spatial
reference is null when the map is paused, such as when the spot report
form is displayed--but only on some devices, which is why I can't
reproduce this issue.

Also got rid of some compiler warnings